+++ src/utils/cloudinary.ts
@@ -6,7 +6,7 @@ import { v4 as uuid } from 'uuid'
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_SECRET
+  api_secret: process.env.CLOUDINARY_SECRET,
 })
 
 /**